+++ CHANGELOG.rst
@@ -1,9 +1,10 @@
 Changelog
 =========
 
-Current (2024-05-24)
+Current (2024-05-28)
 --------------------
 
+* Added I2C address validation function
 * Added endian swap functions to the generic functions file
 
 0.0.2 (2024-05-24)