+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-Current (2024-06-21)
---------------------
+0.1.0 (2024-06-21)
+------------------
 
 * Quickly added address_space_controller, base_chip and etroc2_chip (without associated testing infrastructure) so we can be ready for the next SEU campaign
 * Added additional log levels